requirements.txt
appnope==0.1.0
astroid==1.6.5
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.1
cachetools==2.1.0
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
cssselect==1.0.3
decorator==4.1.2
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
futures==3.2.0
google-api-core==1.4.1
google-auth==1.5.1
google-cloud-pubsub==0.38.0
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.11.0
grpcio-tools==1.11.0
html5lib==1.0b10
idna==2.7
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.0.1
isort==4.3.4
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
lazy-object-proxy==1.3.1
lxml==4.2.5
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.7.4
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.1.0
numpy==1.14.2
pandocfilters==1.4.2
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pyasn1==0.4.4
pyasn1-modules==0.2.2
pydicom==0.9.9
Pygments==2.2.0
pylint==1.9.3
python-dateutil==2.6.1
pytz==2018.5
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.19.1
rsa==4.0
scandir==1.6
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
TBB==0.1
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.3
wrapt==1.10.11